Look at the Map/Chapter 1 Scene I make Slime Enemy!!

+++ Assets/03_Code/Enemy/Interface/IDamageable.cs
@@ -4,7 +4,7 @@
     }
 
     public struct DamageInfo {
-        public float DamageAmount;
+        public int DamageAmount;
         public float KnockbackForce;
     }
